reinforce the public interface
